fix-up build_local.sh|actions

+++ build_local.sh
@@ -16,21 +16,15 @@ export MACOSX_DEPLOYMENT_TARGET=10.14
 # crossover source code to be downloaded
 export CROSS_OVER_SOURCE_URL=https://media.codeweavers.com/pub/crossover/source/crossover-sources-${CROSS_OVER_VERSION}.tar.gz
 export CROSS_OVER_LOCAL_FILE=crossover-${CROSS_OVER_VERSION}
-export CROSS_OVER_LLVM=21.2.0
 # directories / files inside the downloaded tar file directory structure
-export LLVM_MAKEDIR=$GITHUB_WORKSPACE/sources/$([[ ${CROSS_OVER_VERSION} == 2?.* ]] && echo "clang/llvm" || echo "llvm")
-export CLANG_MAKEDIR=$GITHUB_WORKSPACE/sources/$([[ ${CROSS_OVER_VERSION} == 2?.* ]] && echo "clang/clang" || echo "clang")
 export WINE_CONFIGURE=$GITHUB_WORKSPACE/sources/wine/configure
 export DXVK_BUILDSCRIPT=$GITHUB_WORKSPACE/sources/dxvk/package-release.sh
 # build directories
 export BUILDROOT=$GITHUB_WORKSPACE/build
-export LLVM_BUILDDIR=$GITHUB_WORKSPACE/build/llvm
-export CLANG_BUILDDIR=$GITHUB_WORKSPACE/build/clang
 # target directory for installation
 export INSTALLROOT=$GITHUB_WORKSPACE/install
 export PACKAGE_UPLOAD=$GITHUB_WORKSPACE/upload
 # artifact names
-export TOOLS_INSTALLATION=build-tools-cx${CROSS_OVER_VERSION}
 export WINE_INSTALLATION=wine-cx${CROSS_OVER_VERSION}
 export DXVK_INSTALLATION=dxvk-cx${CROSS_OVER_VERSION}
 
@@ -42,58 +36,52 @@ then
     exit
 fi
 
-# Make a wrapper to ensure Intel brew is used and not Apple Silicon brew
-ibrew() {
-    /usr/local/bin/brew "$@"
-}
-export -f ibrew
-
-# Make sure Intel brew Install is before a possible Apple Silicon brew
-export PATH="/usr/local/bin:/usr/local/sbin:${PATH}"
+# Manually configure $PATH
+export PATH="/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:/Library/Apple/usr/bin"
 
 echo Getting latest HomeBrew formulas/bottles
-ibrew update
+brew update
 
 echo Installing Dependencies
 # build dependencies
-ibrew install  bison            \
-               cmake            \
-               mingw-w64        \
+brew install   bison                \
+               cmake                \
+               gcenx/wine/cx-llvm   \
+               mingw-w64            \
                ninja
 
 # runtime dependencies for crossover-wine
-ibrew install  faudio           \
-               freetype         \
-               gnutls           \
-               gphoto2          \
-               gst-plugins-base \
-               libpng           \
-               little-cms2      \
-               molten-vk        \
-               mpg123           \
-               sane-backends    \
+brew install   faudio               \
+               freetype             \
+               gnutls               \
+               gphoto2              \
+               gst-plugins-base     \
+               libpng               \
+               little-cms2          \
+               molten-vk            \
+               mpg123               \
+               sane-backends        \
                sdl2
 
-echo Add bison to PATH
+echo "Add bison to PATH"
 export PATH="$(brew --prefix bison)/bin":${PATH}
 
-echo Add llvm/clang to PATH for later
-export PATH="${INSTALLROOT}/${TOOLS_INSTALLATION}/bin":${PATH}
+echo "Add cx-llvm to PATH"
+export PATH="$(brew --prefix cx-llvm)/bin":${PATH}
 
 
 ############ Download and Prepare Source Code ##############
 
 echo Get Source
-curl -o ${CROSS_OVER_LOCAL_FILE}.tar.gz ${CROSS_OVER_SOURCE_URL}
+if [[ ! -f ${CROSS_OVER_LOCAL_FILE}.tar.gz ]]; then
+    curl -o ${CROSS_OVER_LOCAL_FILE}.tar.gz ${CROSS_OVER_SOURCE_URL}
+fi
 
 echo Extract Source
-tar xf ${CROSS_OVER_LOCAL_FILE}.tar.gz
-
-if [[ "${CROSS_OVER_VERSION}" == "20.0.1" || "${CROSS_OVER_VERSION}" == "20.0.2"  ]]; then
-    echo Add missing llvm/clang
-    curl -o crossover-${CROSS_OVER_LLVM}.tar.gz https://media.codeweavers.com/pub/crossover/source/crossover-sources-${CROSS_OVER_LLVM}.tar.gz
-    tar -xf crossover-${CROSS_OVER_LLVM} sources/clang
+if [[ -d "${GITHUB_WORKSPACE}/sources" ]]; then
+    rm -rf ${GITHUB_WORKSPACE}/sources
 fi
+tar xf ${CROSS_OVER_LOCAL_FILE}.tar.gz
 
 echo "Patch Add missing distversion.h"
 # Patch provided by Josh Dubois, CrossOver product manager, CodeWeavers.
@@ -113,85 +101,41 @@ if [[ ${CROSS_OVER_VERSION} == 20.* ]]; then
     pushd sources/wine
     patch -p1 < ${GITHUB_WORKSPACE}/msvcrt-sincos.patch
     popd
-
-    echo Patch DXVK
-    patch sources/dxvk/src/util/rc/util_rc_ptr.h < dxvk_util_rc_ptr.patch
 fi
-
-############ Build LLVM / Clang ##############
-
-echo Configure LLVM
-mkdir -p ${LLVM_BUILDDIR}
-pushd ${LLVM_BUILDDIR}
-cmake -G Ninja \
-    -DLLVM_TARGETS_TO_BUILD=X86 \
-    -DCMAKE_BUILD_TYPE=Release \
-    -DCMAKE_INSTALL_PREFIX="${INSTALLROOT}/${TOOLS_INSTALLATION}" \
-    ${LLVM_MAKEDIR}
-popd
-
-echo Build LLVM
-pushd ${LLVM_BUILDDIR}
-Ninja
-popd
-
-echo Install LLVM
-pushd ${LLVM_BUILDDIR}
-Ninja install
-popd
-
-echo Configure Clang
-mkdir -p ${CLANG_BUILDDIR}
-pushd ${CLANG_BUILDDIR}
-cmake -G Ninja \
-    -DCMAKE_BUILD_TYPE=Release \
-    -DCMAKE_INSTALL_PREFIX="${INSTALLROOT}/${TOOLS_INSTALLATION}" \
-    ${CLANG_MAKEDIR}
-popd
-
-echo Build Clang
-pushd ${CLANG_BUILDDIR}
-Ninja
-popd
-
-echo Install Clang
-pushd ${CLANG_BUILDDIR}
-Ninja install
-popd
-
-echo Tar Build Tools
-pushd ${INSTALLROOT}
-tar -czf ${TOOLS_INSTALLATION}.tar.gz ${TOOLS_INSTALLATION}
-popd
-
-echo Upload Build Tools
-mkdir -p ${PACKAGE_UPLOAD}
-cp ${INSTALLROOT}/${TOOLS_INSTALLATION}.tar.gz ${PACKAGE_UPLOAD}/
 
 
 ############ Build DXVK ##############
 
-#if [[ ${CROSS_OVER_VERSION} == 20.* ]]; thend
-#    Echo "Installing dependencies for dxvk"
-#    brew install  coreutils \
-#                  meson     \
-#                  glslang
-#    echo Build DXVK
-#    PATH="$(brew --prefix coreutils)/libexec/gnubin:${PATH}" ${DXVK_BUILDSCRIPT} master ${INSTALLROOT}/${DXVK_INSTALLATION} --no-package
-#
-#    echo Tar DXVK
-#    pushd ${INSTALLROOT}
-#    tar -czf ${DXVK_INSTALLATION}.tar.gz ${DXVK_INSTALLATION}
-#    popd
-#
-#    echo Upload DXVK
-#    mkdir -p ${PACKAGE_UPLOAD}
-#    cp ${INSTALLROOT}/${DXVK_INSTALLATION}.tar.gz ${PACKAGE_UPLOAD}/
-#fi
+if [[ ${CROSS_OVER_VERSION} == 21.* ]]; then
+    if [[ ! -f "${PACKAGE_UPLOAD}/${DXVK_INSTALLATION}.tar.gz" ]]; then
+        echo "Applying patches to DXVK"
+        pushd sources/dxvk
+        patch -p1 < ${GITHUB_WORKSPACE}/0001-build-macOS-Fix-up-for-macOS.patch
+        patch -p1 < ${GITHUB_WORKSPACE}/0002-fix-d3d11-header-for-MinGW-9-1883.patch
+        patch -p1 < ${GITHUB_WORKSPACE}/0003-fixes-for-mingw-gcc-12.patch
+        popd
+
+        echo "Installing dependencies for DXVK"
+        brew install  meson     \
+                      glslang
+
+        echo "Build DXVK"
+        ${DXVK_BUILDSCRIPT} master ${INSTALLROOT}/${DXVK_INSTALLATION} --no-package
+
+        echo "Tar DXVK"
+        pushd ${INSTALLROOT}
+        tar -czf ${DXVK_INSTALLATION}.tar.gz ${DXVK_INSTALLATION}
+        popd
+
+        echo "Upload DXVK"
+        mkdir -p ${PACKAGE_UPLOAD}
+        cp ${INSTALLROOT}/${DXVK_INSTALLATION}.tar.gz ${PACKAGE_UPLOAD}/
+    fi
+fi
 
 ############ Build 64bit Version ##############
 
-echo Configure wine64
+echo "Configure wine64-${CROSS_OVER_VERSION}"
 export CC=clang
 export CXX=clang++
 # see https://github.com/Gcenx/macOS_Wine_builds/issues/17#issuecomment-750346843
@@ -204,12 +148,12 @@ export LDFLAGS="-Wl,-headerpad_max_install_names"
 
 export GPHOTO2_CFLAGS="-I$(brew --prefix libgphoto2)/include -I$(brew --prefix libgphoto2)/include/gphoto2"
 export GPHOTO2_PORT_CFLAGS="-I$(brew --prefix libgphoto2)/include -I$(brew --prefix libgphoto2)/include/gphoto2"
-export SDL2_CFLAGS="-I$(brew --prefix sdl2)/"$([[ ${CROSS_OVER_VERSION} -ge 21.* ]] && echo "include/SDL2" || echo "include")
+export SDL2_CFLAGS="-I$(brew --prefix sdl2)/include -I$(brew --prefix sdl2)/include/SDL2"
 export ac_cv_lib_soname_vulkan=""
 export ac_cv_lib_soname_MoltenVK="$(brew --prefix molten-vk)/lib/libMoltenVK.dylib"
 
-mkdir -p ${BUILDROOT}/wine64
-pushd ${BUILDROOT}/wine64
+mkdir -p ${BUILDROOT}/wine64-${CROSS_OVER_VERSION}
+pushd ${BUILDROOT}/wine64-${CROSS_OVER_VERSION}
 ${WINE_CONFIGURE} \
         --disable-option-checking \
         --enable-win64 \
@@ -231,15 +175,15 @@ ${WINE_CONFIGURE} \
         --without-x
 popd
 
-echo Build wine64
-pushd ${BUILDROOT}/wine64
+echo "Build wine64-${CROSS_OVER_VERSION}"
+pushd ${BUILDROOT}/wine64-${CROSS_OVER_VERSION}
 make -j$(sysctl -n hw.ncpu 2>/dev/null)
 popd
 
 
 ############ Build 32bit Version (WoW64) ##############
 
-echo Configure wine32on64
+echo "Configure wine32on64-${CROSS_OVER_VERSION}"
 export CC=clang
 export CXX=clang++
 # see https://github.com/Gcenx/macOS_Wine_builds/issues/17#issuecomment-750346843
@@ -252,14 +196,14 @@ export LDFLAGS="-Wl,-headerpad_max_install_names"
 
 export GPHOTO2_CFLAGS="-I$(brew --prefix libgphoto2)/include -I$(brew --prefix libgphoto2)/include/gphoto2"
 export GPHOTO2_PORT_CFLAGS="-I$(brew --prefix libgphoto2)/include -I$(brew --prefix libgphoto2)/include/gphoto2"
-export SDL2_CFLAGS="-I$(brew --prefix sdl2)/"$([[ ${CROSS_OVER_VERSION} -ge 21.* ]] && echo "include/SDL2" || echo "include")
+export SDL2_CFLAGS="-I$(brew --prefix sdl2)/include -I$(brew --prefix sdl2)/include/SDL2"
 
-mkdir -p ${BUILDROOT}/wine32on64
-pushd ${BUILDROOT}/wine32on64
+mkdir -p ${BUILDROOT}/wine32on64-${CROSS_OVER_VERSION}
+pushd ${BUILDROOT}/wine32on64-${CROSS_OVER_VERSION}
 ${WINE_CONFIGURE} \
         --disable-option-checking \
         --enable-win32on64 \
-        --with-wine64=${BUILDROOT}/wine64 \
+        --with-wine64=${BUILDROOT}/wine64-${CROSS_OVER_VERSION} \
         --disable-tests \
         --without-alsa \
         --without-capi \
@@ -286,32 +230,32 @@ ${WINE_CONFIGURE} \
         --without-x
 popd
 
-echo Build wine32on64
-pushd ${BUILDROOT}/wine32on64
+echo "Build wine32on64-${CROSS_OVER_VERSION}"
+pushd ${BUILDROOT}/wine32on64-${CROSS_OVER_VERSION}
 make -j$(sysctl -n hw.activecpu 2>/dev/null)
 popd
 
 
 ############ Install wine ##############
 
-echo Install wine32on64
-pushd ${BUILDROOT}/wine32on64
+echo "Install wine32on64-${CROSS_OVER_VERSION}"
+pushd ${BUILDROOT}/wine32on64-${CROSS_OVER_VERSION}
 make install-lib DESTDIR="${INSTALLROOT}/${WINE_INSTALLATION}"
 popd
 
-echo Install wine64
-pushd ${BUILDROOT}/wine64
+echo "Install wine64-${CROSS_OVER_VERSION}"
+pushd ${BUILDROOT}/wine64-${CROSS_OVER_VERSION}
 make install-lib DESTDIR="${INSTALLROOT}/${WINE_INSTALLATION}"
 popd
 
 
 ############ Bundle and Upload Deliverable ##############
 
-echo Tar Wine
+echo "Tar Wine"
 pushd ${INSTALLROOT}
 tar -czvf ${WINE_INSTALLATION}.tar.gz ${WINE_INSTALLATION}
 popd
 
-echo Upload Wine
+echo "Upload Wine"
 mkdir -p ${PACKAGE_UPLOAD}
 cp ${INSTALLROOT}/${WINE_INSTALLATION}.tar.gz ${PACKAGE_UPLOAD}/
